+++ src/rds-control-plane-mcp-server/awslabs/rds_control_plane_mcp_server/resources/db_instance/list_performance_reports.py
@@ -16,9 +16,9 @@
 
 import asyncio
 from ...common.connection import PIConnectionManager
-from ...common.exceptions import handle_exceptions
+from ...common.context import RDSContext
+from ...common.decorators.handle_exceptions import handle_exceptions
 from ...common.server import mcp
-from ...context import RDSContext
 from datetime import datetime
 from pydantic import BaseModel, Field
 from typing import List, Literal, Optional
