+++ src/rds-control-plane-mcp-server/awslabs/rds_control_plane_mcp_server/tools/db_instance/change_instance_status.py
@@ -15,17 +15,18 @@
 """Tool to manage the status of an Amazon RDS database instance."""
 
 import asyncio
-from ...common.confirmation import readonly_check, require_confirmation
 from ...common.connection import RDSConnectionManager
-from ...common.exceptions import handle_exceptions
-from ...common.server import mcp
-from ...common.utils import (
-    format_rds_api_response,
-)
-from ...constants import (
+from ...common.constants import (
     SUCCESS_REBOOTED,
     SUCCESS_STARTED,
     SUCCESS_STOPPED,
+)
+from ...common.decorators.handle_exceptions import handle_exceptions
+from ...common.decorators.readonly_check import readonly_check
+from ...common.decorators.require_confirmation import require_confirmation
+from ...common.server import mcp
+from ...common.utils import (
+    format_rds_api_response,
 )
 from loguru import logger
 from mcp.server.fastmcp import Context as FastMCPContext
